my-vapi-agent/requirements.txt
vapi_python==0.1.9
daily-python==0.19.6
pyaudio==0.2.14
requests==2.32.4
python-dotenv==1.1.1
